requirements.txt
asgiref==3.4.1
Django==3.2.8
numpy==1.21.4
pandas==1.3.4
Pillow==8.4.0
PyMySQL==1.0.2
python-dateutil==2.8.2
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
weasyprint==53.4
gunicorn
mysqlclient
sendgrid
requests
django-rest-framework==0.1.0
mailersend==0.2.1
